@@ -25,6 +25,6 @@
     "rollup-plugin-node-resolve": "^5.0.3",
     "rollup-plugin-svelte": "5.1.0",
     "rollup-plugin-terser": "^5.0.0",
-    "svelte": "3.18.1"
+    "svelte": "3.28.0"
   }
 }
