@@ -1,4 +1,4 @@
-import { createSignal, createEffect, createMemo, freeze, sample, mapArray } from 'solid-js';
+import { createSignal, createSelector, batch } from 'solid-js';
 import { render } from 'solid-js/dom';
 
 let idCounter = 1;
@@ -25,24 +25,10 @@ const Button = ({ id, text, fn }) =>
     <button id={ id } class='btn btn-primary btn-block' type='button' onClick={ fn }>{ text }</button>
   </div>
 
-const List = props => {
-  const mapped = createMemo(mapArray(props.each, props.children));
-  createEffect(tr => {
-    let i, s = props.selected();
-    sample(() => {
-      if (tr) tr.className = "";
-      if ((tr = s && (i = props.each().findIndex(el => el.id === s)) > -1 && mapped()[i]))
-        tr.className = "danger";
-    });
-    return tr;
-  });
-  return mapped;
-};
-
 const App = () => {
-  let rowId;
   const [data, setData] = createSignal([]),
-    [selected, setSelected] = createSignal(null, (a, b) => a === b);
+    [selected, setSelected] = createSignal(null, true),
+    isSelected = createSelector(selected);
 
   return <div class='container'>
     <div class='jumbotron'><div class='row'>
@@ -57,15 +43,15 @@ const App = () => {
       </div></div>
     </div></div>
     <table class='table table-hover table-striped test-data'><tbody>
-      <List each={ data } selected={ selected }>{ row => (
-        rowId = row.id,
-        <tr>
+      <For each={ data() }>{ row => {
+        let rowId = row.id;
+        return <tr class={isSelected(rowId) ? "danger": ""}>
           <td class='col-md-1' textContent={ rowId } />
           <td class='col-md-4'><a onClick={[setSelected, rowId]} textContent={ row.label() } /></td>
           <td class='col-md-1'><a onClick={[remove, rowId]}><span class='glyphicon glyphicon-remove' aria-hidden="true" /></a></td>
           <td class='col-md-6'/>
         </tr>
-      )}</List>
+      }}</For>
     </tbody></table>
     <span class='preloadicon glyphicon glyphicon-remove' aria-hidden="true" />
   </div>;
@@ -77,14 +63,14 @@ const App = () => {
   }
 
   function run() {
-    freeze(() => {
+    batch(() => {
       setData(buildData(1000));
       setSelected(null);
     });
   }
 
   function runLots() {
-    freeze(() => {
+    batch(() => {
       setData(buildData(10000));
       setSelected(null);
     });
@@ -93,7 +79,7 @@ const App = () => {
   function add() { setData(data().concat(buildData(1000))); }
 
   function update() {
-    freeze(() => {
+    batch(() => {
       const d = data();
       let index = 0;
       while (index < d.length) {
@@ -114,7 +100,7 @@ const App = () => {
   }
 
   function clear() {
-    freeze(() => {
+    batch(() => {
       setData([]);
       setSelected(null);
     });
